% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmage_funcs.R
\name{get_dynamod_schema}
\alias{get_dynamod_schema}
\title{get_dynamod_schema}
\usage{
get_dynamod_schema()
}
\value{
a pointblack database schema
}
\description{
Returns the pointblank schema required to validate the
dataframe columns which are passed into dynamod.
}
