% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faostat-helpers.R
\name{get_fao_metadata}
\alias{get_fao_metadata}
\title{get_fao_metadata}
\usage{
get_fao_metadata(
  dataset_codes,
  metadata_json = Sys.glob(here::here("data", "FAOSTAT", "*FAOSTAT_datasets_E.json"))
)
}
\arguments{
\item{dataset_codes}{character vector of dataset codes}

\item{metadata_json}{location of the metadata json}
}
\value{
list of dataset metadata
}
\description{
import metadata for set of FAOSTAT datasets
from the FAOSTAT metadata json file
}
\examples{

get_fao_metadata("QCL")

}
