% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmage_funcs.R
\name{run_dynamod_projection}
\alias{run_dynamod_projection}
\title{Runs dynamod truncated projection model}
\usage{
run_dynamod_projection(df, num_phases = 12, num_cycles = 1)
}
\arguments{
\item{df}{data frame containing dynamod parameters}

\item{num_phases}{number of phases within a dynamod cycle defaults to 12}

\item{num_cycles}{number of cycles to run the projections for defaults to 1}
}
\description{
Runs dynamod truncated projection model
}
