% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmage_funcs.R
\name{validate_dynamod_schema}
\alias{validate_dynamod_schema}
\title{validate_dynamod_schema}
\usage{
validate_dynamod_schema(df, .schema = get_dynamod_schema())
}
\arguments{
\item{.schema}{}
}
\value{
logical determining if the dataframe conforms
}
\description{
Validates if the data frame passed into the dynamod functions
has the appropriate schema.
}
